rename crypto module to not clash with imported module of the same name, re #2620
